+++ source/extensions/isaacsim.core.nodes/premake5.lua
@@ -65,6 +65,7 @@ links {
 
 extra_usd_libs = { "usdGeom", "usdPhysics", "ts" }
 
+
 -- Begin OpenUSD
 add_usd(extra_usd_libs)
 -- End OpenUSD